+++ tests/test_model.py
@@ -25,8 +25,10 @@ if __name__ == "__main__":
     dummy_market_features = torch.zeros(100, 16, 63).to(
         "cuda" if torch.cuda.is_available() else "cpu"
     )
-    dummy_industries = torch.zeros(100, 16, 1).to(
-        "cuda" if torch.cuda.is_available() else "cpu"
+    dummy_industries = (
+        torch.zeros(100, 16, 1)
+        .to("cuda" if torch.cuda.is_available() else "cpu")
+        .long()
     )
     output = model.model(dummy_stock_features, dummy_market_features, dummy_industries)
     logger.info(output)
